+++ src/domain/enum/quakeHistrory/earthquakeEnum.ts
