+++ src/domain/enum/quakeHistrory/issueEnum.ts
