+++ src/domain/enum/quakeHistrory/pointsEnum.ts
