+++ src/infrastructure/api/p2pQuake/p2pQuakeApiService.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { HttpService } from '@nestjs/axios';
 import { firstValueFrom } from 'rxjs';
-import { getQuakeHistoryInfoResponseDto } from 'src/application/interfaces/dto/quakeHistoryInfoDto';
+import { getQuakeHistoryInfoResponseDto } from 'src/application/dto/quakeHistoryInfoDto';
 
 /**
  * P2P地震APIサービス
